+++ src/federation/handler.test.ts
@@ -728,6 +728,7 @@ test("handleCollection()", async () => {
   const createCtx = [
     "https://www.w3.org/ns/activitystreams",
     "https://w3id.org/security/data-integrity/v1",
+    "https://w3id.org/identity/v1",
     {
       toot: "http://joinmastodon.org/ns#",
       misskey: "https://misskey-hub.net/ns#",